v3.34.0: MainViewModel partial-class split (Backup + Trello)
Begins breaking up the 820-line MainViewModel god-object identified in the architecture audit. Conservative approach: partial-class extraction (same pattern PlannerService already uses). Public API unchanged, all XAML bindings continue working unmodified, tests untouched.

MainViewModel.Backup.cs (new, ~85 lines):
* BackupDatabase command (SaveFileDialog wrapper around File.Copy)
* RestoreDatabaseAsync command (validate, pre-restore safety copy, replace, reload all VM-loaded collections)

MainViewModel.Trello.cs (new, ~25 lines):
* SaveTrelloSettingsAsync command
* TestTrelloConnectionAsync command
* Trello state (TrelloIsEnabled / TrelloApiKey / etc.) stays in main file because it's the same object shared with settings-page bindings

MainViewModel.cs: 820 -> 623 lines (-197).

Deferred (still candidates for future partials):
* Reminders + Meetings + RecurringTemplates loading commands (~150 lines)
* Update check + install (~50 lines)
* Theme preset application (~50 lines)

This is structural cleanup, not a logical VM split. Going further to per-feature ViewModels (BackupViewModel as a separate object injected into MainViewModel) requires touching XAML bindings and is its own PR.

74 tests pass, dotnet format clean.
